Revertting changes to tile flag to still support -1 but cap at 255 for furture work on #237.

+++ SDK/Engine/Chips/Game/GameChip.cs
@@ -1505,7 +1505,7 @@ namespace PixelVision8.Engine.Chips
         ///     The new value for the flag. Setting the flag to -1 means no collision.
         /// </param>
         /// <returns></returns>
-        public byte Flag(int column, int row, byte? value = null)
+        public int Flag(int column, int row, int? value = null)
         {
             var tile = TilemapChip.GetTile(column, row);
 
@@ -1543,7 +1543,7 @@ namespace PixelVision8.Engine.Chips
         ///     Returns a dictionary containing the spriteID, colorOffset, and flag for an individual tile.
         /// </returns>
         //TODO this should return a custom class not a Dictionary
-        public TileData Tile(int column, int row, int? spriteID = null, int? colorOffset = null, byte? flag = null,
+        public TileData Tile(int column, int row, int? spriteID = null, int? colorOffset = null, int? flag = null,
             bool? flipH = null, bool? flipV = null)
         {
             var invalidateTileMap = false;
@@ -1668,7 +1668,7 @@ namespace PixelVision8.Engine.Chips
         /// <param name="flag">
         ///     An optional flag int value to be applied to each updated tile.
         /// </param>
-        public void UpdateTiles(int[] ids, int? colorOffset = null, byte? flag = null)
+        public void UpdateTiles(int[] ids, int? colorOffset = null, int? flag = null)
         {
             total = ids.Length;
             _width = TilemapSize().X;
